[descriptor] Renamed resource_descriptor files to descriptor.
