+++ src/vulkan-renderer/wrapper/descriptor.cpp
@@ -1,4 +1,4 @@
-#include "inexor/vulkan-renderer/wrapper/resource_descriptor.hpp"
+#include "inexor/vulkan-renderer/wrapper/descriptor.hpp"
 
 #include "inexor/vulkan-renderer/wrapper/device.hpp"
 #include "inexor/vulkan-renderer/wrapper/make_info.hpp"
